fix: Properly set python version for uv (#38)
* chore: bump astral-sh/setup-uv to V5

Cache is now enable by default.
python-version can now be specified to define the UV_PYTHON var for the rest of the workflow.

* ci: have fast tests not fail fast

Fail fast set to true prevents from rerying a failed job in this
workflow if another job failed.

Signed-off-by: Paul Mars <paul.mars@canonical.com>

* fix: Set lowest python version to a valid value in self tests

uv is unable to create a venv with such an old version.
Use 3.8 instead since this below the lower version we support by default in the workflow and
above the older version supported by uv (currently 3.7).

Signed-off-by: Paul Mars <paul.mars@canonical.com>

* fix: do not fail when the uv cache does not exist

In some self-tests there might be nothing to cache for uv. This setting could later be exposed as an input of the workflow if we want to change the default back to false (and thus fail again).

Signed-off-by: Paul Mars <paul.mars@canonical.com>

* fix(scan-python): Do not force generating requirements for all extras

Use uv-sync-extra-args as a multiline string to list groups of cmd line parameters to feed to uv to select the desired extras.

Signed-off-by: Paul Mars <paul.mars@canonical.com>

---------

Signed-off-by: Paul Mars <paul.mars@canonical.com>